// frontend/src/lib/api/orval/api/generated/model/fillBlankExercise.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Langu-AI API
 * API for the Langu-AI German language learning platform
 * OpenAPI spec version: 1.0.0
 */
import type { FillBlankExerciseExerciseType } from "./fillBlankExerciseExerciseType";

export interface FillBlankExercise {
  readonly id: string;
  exercise_type: FillBlankExerciseExerciseType;
  readonly item_id: string;
  readonly item_text: string;
  readonly item_translation: string;
  text_before: string;
  text_after: string;
  answer: string;
  accept_alternatives: string[];
  hint?: string;
  explanation?: string;
}